requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
django-channels-jwt-auth-middleware==1.0.0
django-extensions==3.2.1
django-filter==22.1
django-python3-ldap==0.15.3
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
exceptiongroup==1.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
kafka-python==2.0.2
ldap3==2.9.1
loguru==0.6.0
lxml==4.9.2
msgpack==1.0.4
outcome==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
PySocks==1.7.1
pytz==2022.4
PyYAML==6.0
redis==4.4.0
requests==2.28.1
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.13
websocket-client==1.5.0
websockets==10.3
win32-setctime==1.1.0
wsproto==1.2.0
zope.interface==5.5.0
